requirements.txt
certifi==2018.4.16
chardet==3.0.4
idna==2.7
PyYAML==3.12
requests==2.19.1
six==1.11.0
slackclient==1.2.1
urllib3==1.23
websocket-client==0.48.0
